#133 Pull off Ktor code into separate file
